feat: [dynamic-workflows-cell-website-live-provider-inspection-004] - [Preserve bounded session-detail loading, missing, and error behavior around live-provider details]
